added check if window exists
